requirements.txt
behave==1.2.5
parse==1.6.6
parse-type==0.3.4
PyHamcrest==1.9.0
requests==2.11.1
selenium==3.0.0b3
six==1.10.0
